WIP - pairing discussion.
